ai.video/split-video/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic-settings==2.1.0
openai-whisper==20231117
ffmpeg-python==0.2.0
openai==1.12.0
python-dotenv==1.0.0
aiofiles==23.2.1
